fix a huge bug

+++ src/main/java/com/codimiracle/application/platform/huidu/entity/vo/FigureTagVO.java
@@ -2,10 +2,12 @@ package com.codimiracle.application.platform.huidu.entity.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class FigureTagVO {
     private String id;
     private String userId;
-    private Double score;
+    private BigDecimal score;
     private TagVO tag;
 }
